+++ .changeset/cyan-rules-swim.md
@@ -1,0 +1,5 @@
+---
+'@solana/rpc-transformers': patch
+---
+
+Removed `OPTIONS_OBJECT_POSITION_BY_METHOD`, `downcastNodeToNumberIfBigint()`, `applyDefaultCommitment()`, `getIntegerOverflowNodeVisitor()`, `getBigIntUpcastVisitor()`, and `getTreeWalker()` from the exports of `@solana/rpc-transformer`
