+++ packages/rpc-transformers/src/__tests__/request-transformer-integer-overflow-test.ts
@@ -1,4 +1,4 @@
-import { getIntegerOverflowNodeVisitor } from '../request-transformer-integer-overflow';
+import { getIntegerOverflowNodeVisitor } from '../request-transformer-integer-overflow-internal';
 import { TraversalState } from '../tree-traversal';
 
 const MOCK_TRAVERSAL_STATE = {
