+++ packages/rpc-transformers/src/request-transformer-integer-overflow.ts
@@ -1,6 +1,7 @@
 import { RpcRequest } from '@solana/rpc-spec-types';
 
-import { getTreeWalkerRequestTransformer, KeyPath, TraversalState } from './tree-traversal';
+import { getIntegerOverflowNodeVisitor } from './request-transformer-integer-overflow-internal';
+import { getTreeWalkerRequestTransformer, KeyPath } from './tree-traversal';
 
 export type IntegerOverflowHandler = (request: RpcRequest, keyPath: KeyPath, value: bigint) => void;
 
@@ -11,16 +12,5 @@ export function getIntegerOverflowRequestTransformer(onIntegerOverflow: IntegerO
             { keyPath: [] },
         );
         return transformer(request);
-    };
-}
-
-export function getIntegerOverflowNodeVisitor(onIntegerOverflow: (keyPath: KeyPath, value: bigint) => void) {
-    return <T>(value: T, { keyPath }: TraversalState): T => {
-        if (typeof value === 'bigint') {
-            if (onIntegerOverflow && (value > Number.MAX_SAFE_INTEGER || value < -Number.MAX_SAFE_INTEGER)) {
-                onIntegerOverflow(keyPath as (number | string)[], value);
-            }
-        }
-        return value;
     };
 }
